Separate out pytest functions used by backends
This commit splits apart `nengo.conftest` into the parts that are
used by all backends (now `pytest_nengo.py`) and only by this
repository (now `conftest.py`). `pytest_nengo` is distributed
with setuptools using the `py_modules` argument. `pytest_nengo` is
pointed to by a new entrypoint that informs pytest of these functions
and fixtures that it can load them up before running the test suite.

The --simulator, --simloader, --neurons arguments are now provided
through nengo_simulator, nengo_simloader and nengo_neurons
configuration options. These were almost always added to the
`addopts` list, meaning that they are more appropriate as config
options.

The custom coverage template was also removed. `pytest-cov` should
work correctly now.

Avoid use of global TestConfig by directly accessing pytest configs
