+++ setup.cfg
@@ -48,7 +48,6 @@ max-complexity = 10
 max-line-length = 88
 
 [tool:pytest]
-addopts = -p nengo.tests.options
 xfail_strict = False
 norecursedirs =
     .*
@@ -62,6 +61,7 @@ markers =
     example: Mark a test as an example.
     slow: Mark a test as slow to skip it per default.
 filterwarnings =
+    always
     ignore::ImportWarning
     ignore:(Buffer|Memory|The nengo\.spa):DeprecationWarning
     ignore:Skipping some optimization steps
